requirement.txt
django == 2.2
django-crispy-forms
django-import-export
django-reversion
django-formtools
future
httplib2
six
xlwt
xlsxwriter
requests
pillow
django-pure-pagination
